+++ src/main/java/br/com/infnet/guildaaventureiro/dto/relatorio/RankingParticipacao.java
@@ -1,9 +1,11 @@
 package br.com.infnet.guildaaventureiro.dto.relatorio;
 
+import java.math.BigDecimal;
+
 public record RankingParticipacao(
         String nome,
         Long participacoes,
-        Long recompensaRecebida,
+        BigDecimal recompensaRecebida,
         Long destaquesObtidos
 ) {
 }
